Add Accept-Language negotiation
This is useful to integrate the library
with express apps.

The change makes it so that strong
globalize chooses the best matching language
based on the Accept-Language header of an
HTTP request.

You can call g.http(req), and it will give
you an instance of strong-globalize
already configured for the best-matching
language by comparing the requested
languages with the languages provided
within the intl directory.